refactor(with-x): componentize LogoutButton, CopyIcon, SolanaIcon
