+++ examples/with-x/src/app/dashboard/page.tsx
@@ -1,14 +1,16 @@
 "use client";
 
-import { useState, useEffect } from "react";
+import { useEffect } from "react";
 import Image from "next/image";
 import { useRouter } from "next/navigation";
 import { AuthState, ClientState, useTurnkey } from "@turnkey/react-wallet-kit";
+import { CopyIcon } from "../../components/ui/CopyIcon";
+import { SolanaIcon } from "../../components/ui/SolanaIcon";
+import { LogoutButton } from "../../components/LogoutButton";
 
 export default function Dashboard() {
-  const [copiedField, setCopiedField] = useState<string | null>(null);
   const router = useRouter();
-  const { authState, clientState, user, wallets, logout, createWallet, refreshWallets } = useTurnkey();
+  const { authState, clientState, user, wallets, createWallet, refreshWallets } = useTurnkey();
 
   useEffect(() => {
     if (
@@ -38,21 +40,6 @@ export default function Dashboard() {
     wallets[0]?.accounts?.[0]?.address ??
     "0x";
 
-  const copyToClipboard = async (text: string, fieldName: string) => {
-    try {
-      await navigator.clipboard.writeText(text);
-      setCopiedField(fieldName);
-      setTimeout(() => setCopiedField(null), 2000);
-    } catch (err) {
-      console.error("Failed to copy text: ", err);
-    }
-  };
-
-  const handleLogout = async () => {
-    await logout();
-    router.push("/");
-  };
-
   return (
     <main className="min-h-screen bg-background p-6">
       <div className="max-w-2xl mx-auto">
@@ -80,39 +67,7 @@ export default function Dashboard() {
                 </label>
                 <p className="text-foreground font-mono">{userId}</p>
               </div>
-              <button
-                onClick={() => copyToClipboard(userId, "userId")}
-                className="p-2 hover:bg-muted rounded transition-colors"
-                title="Copy User ID"
-              >
-                {copiedField === "userId" ? (
-                  <svg
-                    className="w-4 h-4 text-green-600"
-                    fill="currentColor"
-                    viewBox="0 0 20 20"
-                  >
-                    <path
-                      fillRule="evenodd"
-                      d="M16.707 5.293a1 1 0 010 1.414l-8 8a1 1 0 01-1.414 0l-4-4a1 1 0 011.414-1.414L8 12.586l7.293-7.293a1 1 0 011.414 0z"
-                      clipRule="evenodd"
-                    />
-                  </svg>
-                ) : (
-                  <svg
-                    className="w-4 h-4 text-muted-foreground"
-                    fill="none"
-                    stroke="currentColor"
-                    viewBox="0 0 24 24"
-                  >
-                    <path
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      strokeWidth={2}
-                      d="M8 16H6a2 2 0 01-2-2V6a2 2 0 012-2h8a2 2 0 012 2v2m-6 12h8a2 2 0 002-2v-8a2 2 0 00-2-2h-8a2 2 0 00-2 2v8a2 2 0 002 2z"
-                    />
-                  </svg>
-                )}
-              </button>
+              <CopyIcon text={userId} title="Copy User ID" />
             </div>
           </div>
 
@@ -122,31 +77,7 @@ export default function Dashboard() {
             </h2>
             <div className="flex items-center justify-between p-4 bg-muted/50 rounded-lg">
               <div className="flex items-center gap-3">
-                <svg className="w-6 h-6" viewBox="0 0 397.7 311.7" fill="none">
-                  <linearGradient
-                    id="solanaGradient"
-                    x1="360.8791"
-                    y1="351.4553"
-                    x2="141.213"
-                    y2="-69.2936"
-                    gradientUnits="userSpaceOnUse"
-                  >
-                    <stop offset="0" stopColor="#00FFA3" />
-                    <stop offset="1" stopColor="#DC1FFF" />
-                  </linearGradient>
-                  <path
-                    d="M64.6 237.9c2.4-2.4 5.7-3.8 9.2-3.8h317.4c5.8 0 8.7 7 4.6 11.1l-62.7 62.7c-2.4 2.4-5.7 3.8-9.2 3.8H6.5c-5.8 0-8.7-7-4.6-11.1L64.6 237.9z"
-                    fill="url(#solanaGradient)"
-                  />
-                  <path
-                    d="M64.6 3.8C67.1 1.4 70.4 0 73.8 0h317.4c5.8 0 8.7 7 4.6 11.1l-62.7 62.7c-2.4 2.4-5.7 3.8-9.2 3.8H6.5c-5.8 0-8.7-7-4.6-11.1L64.6 3.8z"
-                    fill="url(#solanaGradient)"
-                  />
-                  <path
-                    d="M333.1 120.1c-2.4-2.4-5.7-3.8-9.2-3.8H6.5c-5.8 0-8.7 7-4.6 11.1l62.7 62.7c2.4 2.4 5.7 3.8 9.2 3.8h317.4c5.8 0 8.7-7 4.6-11.1l-62.7-62.7z"
-                    fill="url(#solanaGradient)"
-                  />
-                </svg>
+                <SolanaIcon />
                 <div>
                   <p className="text-foreground font-medium">{walletName}</p>
                   <p className="text-sm text-muted-foreground">Solana Wallet</p>
@@ -156,50 +87,14 @@ export default function Dashboard() {
                 <span className="text-foreground font-mono text-sm truncate overflow-hidden max-w-[200px]">
                   {walletAddress}
                 </span>
-                <button
-                  onClick={() => copyToClipboard(walletAddress, "walletId")}
-                  className="p-2 hover:bg-muted rounded transition-colors"
-                  title="Copy Wallet ID"
-                >
-                  {copiedField === "walletId" ? (
-                    <svg
-                      className="w-4 h-4 text-green-600"
-                      fill="currentColor"
-                      viewBox="0 0 20 20"
-                    >
-                      <path
-                        fillRule="evenodd"
-                        d="M16.707 5.293a1 1 0 010 1.414l-8 8a1 1 0 01-1.414 0l-4-4a1 1 0 011.414-1.414L8 12.586l7.293-7.293a1 1 0 011.414 0z"
-                        clipRule="evenodd"
-                      />
-                    </svg>
-                  ) : (
-                    <svg
-                      className="w-4 h-4 text-muted-foreground"
-                      fill="none"
-                      stroke="currentColor"
-                      viewBox="0 0 24 24"
-                    >
-                      <path
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        strokeWidth={2}
-                        d="M8 16H6a2 2 0 01-2-2V6a2 2 0 012-2h8a2 2 0 012 2v2m-6 12h8a2 2 0 002-2v-8a2 2 0 00-2-2h-8a2 2 0 00-2 2v8a2 2 0 002 2z"
-                      />
-                    </svg>
-                  )}
-                </button>
+                <CopyIcon text={walletAddress} title="Copy Wallet Address" />
               </div>
             </div>
           </div>
         </div>
+
         <div className="mt-8 flex justify-center">
-          <button
-            onClick={handleLogout}
-            className="px-6 py-2 bg-red-600 hover:bg-red-700 text-white rounded-lg transition-colors font-medium"
-          >
-            Log Out
-          </button>
+          <LogoutButton />
         </div>
       </div>
     </main>
