+++ examples/with-x/src/components/ui/CopyIcon.tsx
@@ -1,0 +1,52 @@
+"use client";
+
+import { useState } from "react";
+
+export function CopyIcon({ text, title }: { text: string; title?: string }) {
+  const [copied, setCopied] = useState(false);
+
+  const handleCopy = async () => {
+    try {
+      await navigator.clipboard.writeText(text);
+      setCopied(true);
+      setTimeout(() => setCopied(false), 2000);
+    } catch (err) {
+      console.error("Failed to copy text: ", err);
+    }
+  };
+
+  if (copied) {
+    return (
+      <svg
+        className="w-4 h-4 text-green-600 cursor-pointer"
+        fill="currentColor"
+        viewBox="0 0 20 20"
+        title={title}
+      >
+        <path
+          fillRule="evenodd"
+          d="M16.707 5.293a1 1 0 010 1.414l-8 8a1 1 0 01-1.414 0l-4-4a1 1 0 011.414-1.414L8 12.586l7.293-7.293a1 1 0 011.414 0z"
+          clipRule="evenodd"
+        />
+      </svg>
+    );
+  }
+
+  return (
+    <svg
+      className="w-4 h-4 text-muted-foreground cursor-pointer hover:text-foreground transition-colors"
+      fill="none"
+      stroke="currentColor"
+      viewBox="0 0 24 24"
+      onClick={handleCopy}
+      title={title}
+    >
+      <path
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        strokeWidth={2}
+        d="M8 16H6a2 2 0 01-2-2V6a2 2 0 012-2h8a2 2 0 012 2v2m-6 12h8a2 2 0 002-2v-8a2 2 0 00-2-2h-8a2 2 0 00-2 2v8a2 2 0 002 2z"
+      />
+    </svg>
+  );
+}
