+++ examples/with-x/src/components/ui/SolanaIcon.tsx
@@ -1,0 +1,29 @@
+export function SolanaIcon({ className = "w-6 h-6" }: { className?: string }) {
+  return (
+    <svg className={className} viewBox="0 0 397.7 311.7" fill="none">
+      <linearGradient
+        id="solanaGradient"
+        x1="360.8791"
+        y1="351.4553"
+        x2="141.213"
+        y2="-69.2936"
+        gradientUnits="userSpaceOnUse"
+      >
+        <stop offset="0" stopColor="#00FFA3" />
+        <stop offset="1" stopColor="#DC1FFF" />
+      </linearGradient>
+      <path
+        d="M64.6 237.9c2.4-2.4 5.7-3.8 9.2-3.8h317.4c5.8 0 8.7 7 4.6 11.1l-62.7 62.7c-2.4 2.4-5.7 3.8-9.2 3.8H6.5c-5.8 0-8.7-7-4.6-11.1L64.6 237.9z"
+        fill="url(#solanaGradient)"
+      />
+      <path
+        d="M64.6 3.8C67.1 1.4 70.4 0 73.8 0h317.4c5.8 0 8.7 7 4.6 11.1l-62.7 62.7c-2.4 2.4-5.7 3.8-9.2 3.8H6.5c-5.8 0-8.7-7-4.6-11.1L64.6 3.8z"
+        fill="url(#solanaGradient)"
+      />
+      <path
+        d="M333.1 120.1c-2.4-2.4-5.7-3.8-9.2-3.8H6.5c-5.8 0-8.7 7-4.6 11.1l62.7 62.7c2.4 2.4 5.7 3.8 9.2 3.8h317.4c5.8 0 8.7-7 4.6-11.1l-62.7-62.7z"
+        fill="url(#solanaGradient)"
+      />
+    </svg>
+  );
+}
